test: Move `aws-request` to global utils
